+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed an issue where links couldn't be opened from the Home tab.
 - Fix delay in results immediately after opening the discover tab.
 - Fixed the 3d card effect on the Notifications tab.
+- Updated layout for search results and mention autocomplete cards.
 
 ## [0.1 (93)] - 2023-11-10Z